--- web/packages/new/photos/services/search.ts
@@ -1,13 +1,0 @@
-import type { EnteFile } from "@/new/photos/types/file";
-
-/**
- * A massaged version of {@link CGroup} suitable for being shown in search
- * results.
- */
-export interface SearchPerson {
-    id: string;
-    name?: string;
-    files: number[];
-    displayFaceID: string;
-    displayFaceFile: EnteFile;
-}
